style fixes for combined inputs
