+++ src/lib/NodeRenderer.js
@@ -10,6 +10,7 @@ import Footer from "./Footer";
 import { slugify } from "./utils/helpers";
 import Select from "react-select";
 import { useEffect, useState } from "react";
+import ManywaysSelect from "./CustomInputs/ManywaysSelect";
 
 const isFormWithOneChoiceFieldOnly = (formSchema, uiSchema) => {
   if (!!formSchema?.properties) {
@@ -31,74 +32,7 @@ const isFormWithOneChoiceFieldOnly = (formSchema, uiSchema) => {
   return false;
 };
 
-const selectStyles = {
-  clearIndicator: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  container: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  control: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  dropdownIndicator: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  group: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  groupHeading: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  indicatorsContainer: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  indicatorSeparator: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  input: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  loadingIndicator: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  loadingMessage: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  menu: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  menuList: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  menuPortal: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  multiValue: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  multiValueLabel: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  multiValueRemove: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  noOptionsMessage: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  option: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  placeholder: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  singleValue: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-  valueContainer: (baseStyles, state) => {
-    return { ...baseStyles };
-  },
-};
+
 
 const NodeRenderer = (props) => {
   const {
@@ -204,103 +138,7 @@ const NodeRenderer = (props) => {
                 RadioWidget: ManywaysRadioWidget,
                 CheckboxesWidget: ManywaysCheckboxWidget,
                 SelectWidget: ManywaysSelectWidget,
-                Select: ({ value, onChange, disabled, ...props }) => {
-                  const { options } = props;
-                  const [menuIsOpen, setMenuIsOpen] = useState(false);
-
-                  useEffect(() => {
-                    const _theOptions =
-                      !!options?.enumOptions &&
-                      !!options?.enumOptions?.length > 0
-                        ? options?.enumOptions
-                        : temp_opts;
-                    if (_theOptions?.length === 1) {
-                      onChange(_theOptions[0].value);
-                    }
-                  }, []);
-
-                  let temp_opts = [
-                    { value: "xxx", label: "XXX" },
-                    { value: "Alberta", label: "Alberta" },
-                  ];
-                  const theOptions =
-                    !!options?.enumOptions && !!options?.enumOptions?.length > 0
-                      ? options?.enumOptions
-                      : temp_opts;
-                  return (
-                    <>
-                      {!!menuIsOpen && (
-                        <>
-                          <div
-                            style={{
-                              position: "fixed",
-                              top: "0",
-                              left: "0",
-                              width: "100%",
-                              height: "100%",
-                              zIndex: "1",
-                              backgroundColor: "red",
-                              opacity: "0",
-                            }}
-                            onClick={(e) => {
-                              setMenuIsOpen(false);
-                            }}
-                          ></div>
-                          <div
-                            style={{
-                              position: "absolute",
-                              top: "0",
-                              right: "0",
-                              width: "80px",
-                              height: "100%",
-                              zIndex: "20",
-                              backgroundColor: "red",
-                              opacity: "0",
-                              cursor: "pointer",
-                            }}
-                            onClick={(e) => {
-                              setMenuIsOpen(false);
-                            }}
-                          ></div>
-                        </>
-                      )}
-                      <Select
-                        onChange={(v) => {
-                          console.log(v);
-                          onChange(v.value);
-                          setMenuIsOpen(false);
-                        }}
-                        onMenuOpen={() => {
-                          // onChange(null);
-                          // console.log(
-                          //   document
-                          //     .querySelector("manyways-wrapper")
-                          //     .shadowRoot.getElementById("react-select-2-listbox")
-                          // );
-                        }}
-                        onFocus={() => {
-                          console.log("focus");
-                          setMenuIsOpen(true);
-                        }}
-                        onDropdownClose={() => {
-                          console.log("dd close");
-                        }}
-                        blurInputOnSelect={true}
-                        onBlur={() => {
-                          console.log("blur");
-                          // setMenuIsOpen(false);
-                        }}
-                        menuIsOpen={menuIsOpen}
-                        styles={selectStyles}
-                        isDisabled={disabled}
-                        value={theOptions.find((o) => o.value === value)}
-                        placeholder={props.placeholder}
-                        options={theOptions}
-                        classNamePrefix="select-mw"
-                      />
-                    </>
-                  );
-                },
+                Select: ManywaysSelect
               }}
               onChange={(e) => {
                 if (!!singleChoiceField) {
